Moving MultiDeviceIterator from contrib to core.
PiperOrigin-RevId: 214173896
